+++ src/cqterrain/damage/blast.py
@@ -3,13 +3,13 @@ from cadqueryhelper import shape
 import random
 
 def random_nudge_points(
-        seed, 
+        seed:str, 
         points, 
-        x_jiggle = (-2,8), 
-        y_jiggle = (-1,1)
-    ):
+        x_jiggle:tuple[int,int]|int = (-2,8), 
+        y_jiggle:tuple[int,int]|int = (-1,1)
+    ) -> list[tuple[int,int]]:
     random.seed(seed)
-    mod_points = []
+    mod_points:list[tuple[int,int]] = []
     for point in points:
         if type(x_jiggle) is tuple:
             r_x = random.randrange(x_jiggle[0], x_jiggle[1])
@@ -27,21 +27,21 @@ def random_nudge_points(
     return mod_points
 
 def blast(
-        seed="test",
-        height = 3,
-        count = (5,10),
-        x_jiggle = (-2,8), 
-        y_jiggle = (-1,1),
-        ring_params = [
+        seed:str = "test",
+        height:float = 3,
+        count:tuple[int,int]|int = (5,10),
+        x_jiggle:tuple[int,int]|int = (-2,8), 
+        y_jiggle:tuple[int,int]|int = (-1,1),
+        ring_params:list[dict] = [
             {"radius": 30, "start_angle":0}, 
             {"radius":20,"start_angle":8}
         ]
-    ):
+    ) -> cq.Workplane:
     random.seed(seed)
-    point_count = count
+    point_count:int = count #type:ignore
     
     if type(point_count) is tuple:
-        point_count = random.randrange(count[0], count[1])
+        point_count = random.randrange(count[0], count[1]) #type:ignore
     
     point_lists = []
     for ring_param in ring_params: