+++ src/cqterrain/greeble/__init__.py
@@ -1,1 +1,2 @@
 from .vent import vent
+from .spokedWheel import spoked_wheel