+++ src/cqterrain/pipe/corrugated.py
@@ -1,18 +1,26 @@
 import cadquery as cq
 import math
 
-def __make_pipe(length, radius):
+def __make_pipe(
+        length:float, 
+        radius:float
+    ) -> cq.Workplane:
     inner_cable = (
         cq.Workplane("XY")
         .cylinder(length, radius)
     ).rotate((0,1,0),(0,0,0),90)
     return inner_cable
 
-def __make_segments(length, segment_length, space, segment):
+def __make_segments(
+        length:float, 
+        segment_length:float, 
+        space:float, 
+        segment:cq.Workplane
+    ) -> cq.Workplane:
     count = math.floor(length / (segment_length + space))
 
-    def __add_segment(loc):
-        return segment.val().located(loc)
+    def __add_segment(loc:cq.Location) -> cq.Shape:
+        return segment.val().located(loc) #type:ignore
 
     if not count:
         count = 1
@@ -31,12 +39,12 @@ def __make_segments(length, segment_length, space, segment):
     return segments
 
 def corrugated_straight(
-    length = 50,
-    radius = 5,
-    inner_radius = 3,
-    segment_length = 5,
-    space = 5
-):    
+    length:float = 50,
+    radius:float = 5,
+    inner_radius:float = 3,
+    segment_length:float = 5,
+    space:float = 5
+) -> cq.Workplane:    
     inner_pipe = __make_pipe(length,inner_radius)
     segment = __make_pipe(segment_length, radius)
     segments = __make_segments(length, segment_length, space, segment)