+++ src/cqterrain/tile/conduit.py
@@ -3,20 +3,20 @@ from .. import pipe as terrain_pipe
 import math
 
 def conduit(
-        length = 25, # length of the tile
-        width = 25, # with of the tile
-        height = 4, # height of the tile
-        frame= 1, # size of the frame can be 0
-        frame_depth =3, # depth of the frame that the pipe is set into, can be zero
-        pipe_count = None, # hard coded pipe count. If falsy the pipes count will be determined by the witdth of the tile and the diameter of the pipe.
+        length:float = 25, # length of the tile
+        width:float = 25, # with of the tile
+        height:float = 4, # height of the tile
+        frame:float = 1, # size of the frame can be 0
+        frame_depth:float =3, # depth of the frame that the pipe is set into, can be zero
+        pipe_count:int|None = None, # hard coded pipe count. If falsy the pipes count will be determined by the witdth of the tile and the diameter of the pipe.
         
-        radius = 4, # radius of the pipe
-        inner_radius = 3, # internal radius of the inner pope
-        segment_length = 6, # size of the pipe segments
-        space = 4, # space between the pipe segments
+        radius:float = 4, # radius of the pipe
+        inner_radius:float = 3, # internal radius of the inner pope
+        segment_length:float = 6, # size of the pipe segments
+        space:float = 4, # space between the pipe segments
         
-        pipe_padding = 1 # padding between pipes
-    ):
+        pipe_padding:float = 1 # padding between pipes
+    ) -> cq.Workplane:
     
     tile = cq.Workplane("XY")
     outline = cq.Workplane("XY").box(length, width, height)
@@ -49,8 +49,8 @@ def conduit(
         
     # if pipe_count is still falsy skip adding pipes altogether.
     if pipe_count:
-        def __add_pipe(loc):
-            return half_pipe.val().located(loc)
+        def __add_pipe(loc:cq.Location)->cq.Shape:
+            return half_pipe.val().located(loc) #type:ignore
         
         half_pipes = (
             cq.Workplane("XY")