+++ changes.md
@@ -3,6 +3,7 @@
 ## 1.0.0
 * Split window.py file out into it's own module with each function being it's own file.
 * Renamed out directory to stl
+* moved spokedWheel from tile to greeble module
 
 ## 0.3.3
 * Upped cadqueryhelper version 0.2.1